+++ src/app/app.js
@@ -2,8 +2,23 @@
 
 var wsNinja = angular.module('app', [
   'ui.router', 'ngResource', 'ngAnimate',
-  'angular-loading-bar',
+  'ngSanitize', 'angular-loading-bar',
 ]);
+
+wsNinja.run(
+  [
+    '$rootScope', '$state',
+    function($rootScope, $state) {
+      $rootScope.$state = $state;
+      $rootScope.$on('$routeChangeSuccess', function(event, data) {
+        data.$promise.then(function(res) {
+          $rootScope.$state.current.data.pageMeta = res.meta;
+        });
+      });
+    },
+
+  ]
+);
 
 wsNinja.config(function(cfpLoadingBarProvider) {
   cfpLoadingBarProvider.includeSpinner = true;
@@ -11,6 +26,6 @@ wsNinja.config(function(cfpLoadingBarProvider) {
 
 wsNinja.controller('NavCtrl', function($scope, $state) {
   $scope.isActive = function(state) {
-    return state === $state.current.name;
+    return $state.current.name.indexOf(state) !== -1;
   };
 });