+++ src/app/controllers/postsCtrl.js
@@ -6,4 +6,9 @@
     $scope.posts = Posts.list();
   });
 
+  wsNinja.controller('PostDetailsCtrl', function($scope, $stateParams, Posts) {
+    $scope.post = Posts.find({id: $stateParams.id});
+    $scope.$emit('$routeChangeSuccess', $scope.post);
+  });
+
 })();