+++ src/app/services/posts.js
@@ -5,13 +5,21 @@
 
   wsNinja.service('Posts', function($resource) {
     return $resource(
-      '/content/posts-list.json',
-      {},
+      '/content/posts/:id.json',
+      {id: '@id'},
       {
         list: {
           method: 'GET',
           isArray: true,
           url: '/content/posts-list.json',
+          transformResponse: function(data) {
+            return angular.fromJson(data);
+          },
+        },
+        find: {
+          method: 'GET',
+          isArray: false,
+          url: '/content/posts/:id.json',
           transformResponse: function(data) {
             return angular.fromJson(data);
           },